test: Split the test target
The current scenario is All the test cases are merged in one fat
exectable. It make it easy to run them but it take too long to build or
run. The idea from spliting the test for now to make it fast.
Sourcetrail_test contains most of small tests. Sourcetrail_func
contains the long runtime test cases.

ticket: SOUR-5